requirements.txt
numpy==1.24.4
Pillow==10.0.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.50.2
